changed code as requested
