--- core/src/com/mygdx/insanerpg/item/Item.java
@@ -1,10 +1,0 @@
-package com.mygdx.insanerpg.item;
-
-public class Item {
-    private int _sellPrice;
-    private int _buyPrice;
-    private String _name;
-    private String _rarity;
-    private int _itemID;
-}
-
