--- core/src/com/mygdx/insanerpg/item/equipment/Equipment.java
@@ -1,7 +1,0 @@
-package com.mygdx.insanerpg.item.equipment;
-
-import com.mygdx.insanerpg.item.Item;
-
-public class Equipment extends Item {
-    private int _durability;
-}
