--- core/src/com/mygdx/insanerpg/item/equipment/armor/Armor.java
@@ -1,6 +1,0 @@
-package com.mygdx.insanerpg.item.equipment.armor;
-
-import com.mygdx.insanerpg.item.equipment.Equipment;
-
-public class Armor extends Equipment {
-}
