--- core/src/com/mygdx/insanerpg/item/equipment/weapon/Scythe.java
@@ -1,4 +1,0 @@
-package com.mygdx.insanerpg.item.equipment.weapon;
-
-public class Scythe extends Weapon{
-}
