--- core/src/com/mygdx/insanerpg/item/equipment/weapon/Sword.java
@@ -1,6 +1,0 @@
-package com.mygdx.insanerpg.item.equipment.weapon;
-
-
-public class Sword extends Weapon{
-
-}
