--- core/src/com/mygdx/insanerpg/item/equipment/weapon/Weapon.java
@@ -1,11 +1,0 @@
-package com.mygdx.insanerpg.item.equipment.weapon;
-
-import com.mygdx.insanerpg.item.equipment.Equipment;
-
-public class Weapon extends Equipment {
-    private int _damage;
-    private int _attackspeed;
-    private int _range;
-    private boolean _isOneHanded;
-
-}
