--- core/src/com/mygdx/insanerpg/item/stackable/Stackable.java
@@ -1,6 +1,0 @@
-package com.mygdx.insanerpg.item.stackable;
-
-import com.mygdx.insanerpg.item.Item;
-
-public class Stackable extends Item {
-}
