--- core/src/com/mygdx/insanerpg/item/stackable/consumable/Consumable.java
@@ -1,9 +1,0 @@
-package com.mygdx.insanerpg.item.stackable.consumable;
-
-import com.mygdx.insanerpg.item.Item;
-
-public class Consumable extends Item {
-    private int _cooldown;
-    private int _hasDelay;
-    private int _delay;
-}
