--- core/src/com/mygdx/insanerpg/item/stackable/consumable/Potion.java
@@ -1,5 +1,0 @@
-package com.mygdx.insanerpg.item.stackable.consumable;
-
-public class Potion extends Consumable {
-
-}
