+++ core/src/com/mygdx/insanerpg/model/item/Item.java
@@ -1,0 +1,8 @@
+package com.mygdx.insanerpg.model.item;
+
+public class Item {
+    private int _sellPrice;
+    private int _buyPrice;
+    private int _itemId;
+}
+
