+++ core/src/com/mygdx/insanerpg/model/item/equipment/Equipment.java
@@ -1,0 +1,8 @@
+package com.mygdx.insanerpg.model.item.equipment;
+
+import com.mygdx.insanerpg.model.item.Item;
+
+public class Equipment extends Item {
+    private int _maxDurability;
+    private int _currentDurability;
+}
