+++ core/src/com/mygdx/insanerpg/model/item/equipment/armor/Armor.java
@@ -1,0 +1,6 @@
+package com.mygdx.insanerpg.model.item.equipment.armor;
+
+import com.mygdx.insanerpg.model.item.equipment.Equipment;
+
+public class Armor extends Equipment {
+}
