+++ core/src/com/mygdx/insanerpg/model/item/equipment/weapon/Scythe.java
@@ -1,0 +1,4 @@
+package com.mygdx.insanerpg.model.item.equipment.weapon;
+
+public class Scythe extends Weapon{
+}
