+++ core/src/com/mygdx/insanerpg/model/item/equipment/weapon/Sword.java
@@ -1,0 +1,5 @@
+package com.mygdx.insanerpg.model.item.equipment.weapon;
+
+
+public class Sword extends Weapon{
+}
