+++ core/src/com/mygdx/insanerpg/model/item/equipment/weapon/Weapon.java
@@ -1,0 +1,10 @@
+package com.mygdx.insanerpg.model.item.equipment.weapon;
+
+import com.mygdx.insanerpg.model.item.equipment.Equipment;
+
+public class Weapon extends Equipment {
+    private int _basePhysicalDamage;
+    private int _baseMagicalDamage;
+    private boolean _isOneHanded;
+
+}
