+++ core/src/com/mygdx/insanerpg/model/item/stackable/Stackable.java
@@ -1,0 +1,6 @@
+package com.mygdx.insanerpg.model.item.stackable;
+
+import com.mygdx.insanerpg.model.item.Item;
+
+public class Stackable extends Item {
+}
