+++ core/src/com/mygdx/insanerpg/model/item/stackable/consumable/Consumable.java
@@ -1,0 +1,10 @@
+package com.mygdx.insanerpg.model.item.stackable.consumable;
+
+import com.mygdx.insanerpg.model.item.Item;
+
+public class Consumable extends Item {
+    private int _cooldown;
+    private int _hasDelay;
+    private int _delay;
+    private int _duration;
+}
