+++ core/src/com/mygdx/insanerpg/model/item/stackable/consumable/Potion.java
@@ -1,0 +1,5 @@
+package com.mygdx.insanerpg.model.item.stackable.consumable;
+
+public class Potion extends Consumable {
+
+}
